# Optimize `cmp` function.