Before:
% time ./v7 fib.js
[0, 1, 1, 2, 3, 5, 8, 13, 21, 34, 55, 89, 144, 233, 377, 610, 987, 1597,
2584, 4181, 6765, 10946, 17711, 28657, 46368, 75025, 121393, 196418,
317811, 514229, 832040]
7049123
./v7 fib.js  71,47s user 0,13s system 99% cpu 1:11,85 total

After:
% time ./v7 fib.js
[0, 1, 1, 2, 3, 5, 8, 13, 21, 34, 55, 89, 144, 233, 377, 610, 987, 1597,
2584, 4181, 6765, 10946, 17711, 28657, 46368, 75025, 121393, 196418,
317811, 514229, 832040]
7049123
./v7 fib.js  42,39s user 0,08s system 99% cpu 42,573 total

Compiled with gcc 4.8 using

% gcc -O3 -g -DNDEBUG -DV7_EXE v7.c -o v7 -lm