@cpq 2015-01-02
v7.c is a generated file, amalgamated from src/*.c
Please update the source, src/util.c then do `make` in the root level dir to generate v7.c