@vstakhov 2015-01-02
Done. However, v7 is still terribly slow. I have the updated profile for this simple test. Previously, the function `cmp` ate about 40% of time. At the moment, there are no such specific functions, however the overall performance is still unacceptable. 30 Fibonacci numbers means 7M of recursive calls approximately: `v8` performs that in likely 0.15 seconds, `duktape` - in 4.5 seconds (which is awful as well) and v7 spends about 40 seconds now (but it is definitely better than 72 seconds without this patch).