@cpq 2015-01-03
I still don't see changes in `util.c`. Did you do `git commit -a --amend` followed by `git push -f` ?