@cpq 2015-01-03
Ah, got it. There was a second commit.
Merging.
W.r.t. performance - agree with the point. We're in the middle of major core refactoring, which is going to improve things a lot. The 4 loosely coupled components of the engines are: tokenizer, parser, ast, VM - and first three are more or less done. We can parse all ECMA tests and libs like jquery, and generate AST for those. Work is in progress on VM part.
Note that right now an "old" components coexists with "new" ones: like, old "parser.c" with new "aparser.c".